+++ src/rest.ts
@@ -2,6 +2,8 @@ export {ResponseType, LocationHeaderAndJsonResponse, LocationHeaderResponse, Blo
 export {Cache} from './rest/cache';
 export {DELETE, GET, HEAD, POST, PUT, RESTClient} from './rest/common';
 export {RestTransferStateService} from './transferState/restTransferState.service';
-export {AdditionalInfoPropertyDescriptor, ParametersMetadata, RestCaching, RestDisabledInterceptors, RestFullHttpResponse, RestHttpHeaders, RestMethod, RestParameters, RestReportProgress, RestResponseTransform, RestResponseType, KeyIndex, ParametersTransformMetadata, RestMiddleware} from './rest/rest.interface';
+export {AdditionalInfoPropertyDescriptor, ParametersMetadata, RestCaching, RestDisabledInterceptors, RestFullHttpResponse, RestHttpHeaders, RestMethod, RestParameters, RestReportProgress, RestResponseTransform, RestResponseType, KeyIndex, ParametersTransformMetadata, RestMiddleware, RestMiddlewareRunMethod, ɵRESTClient, RestMethodMiddlewares} from './rest/rest.interface';
 export {HTTP_HEADER_ACCEPT, HTTP_HEADER_CONTENT_TYPE} from './rest/constants';
 export {Body, BaseUrl, DefaultHeaders, Header, Headers, JsonContentType, ParameterTransform, Path, Produces, Query, QueryObject, ResponseTransform, FullHttpResponse, ReportProgress, DisableInterceptor, ProgressIndicatorGroup} from './rest/decorators';
+export {REST_METHOD_MIDDLEWARES, REST_MIDDLEWARES_ORDER} from './rest/tokens';
+export {BASIC_DEFAULT_REST_METHOD_MIDDLEWARES, BASIC_DEFAULT_REST_MIDDLEWARES_ORDER} from './rest/defaults';