+++ src/rest/defaults.ts
@@ -1,0 +1,17 @@
+import {Type} from '@angular/core';
+
+import {RestMiddleware} from './rest.interface';
+
+/**
+ * Definition of basic default array of rest middlewares order
+ */
+export const BASIC_DEFAULT_REST_MIDDLEWARES_ORDER: Type<RestMiddleware>[] =
+[
+];
+
+/**
+ * Definition of basic defaut array of rest middlewares used for each rest method
+ */
+export const BASIC_DEFAULT_REST_METHOD_MIDDLEWARES: Type<RestMiddleware>[] =
+[
+];