+++ src/rest/utils.ts
@@ -1,15 +1,13 @@
 import {Type} from '@angular/core';
 
-import {RESTClientInterface, RestMiddleware, RestMiddlewareRunMethod} from './rest.interface';
+import {RestMiddleware, RestMiddlewareRunMethod} from './rest.interface';
 
 /**
  * Builds and returns array of middleware run functions
- * @param this - Method should be bound to RESTClient instance
  * @param middlewares - Array of set middleware types
  * @param middlewaresOrder - Array of middleware types in order that should be executed
  */
-export function buildMiddlewares(this: RESTClientInterface,
-                                 middlewares: Type<RestMiddleware>[],
+export function buildMiddlewares(middlewares: Type<RestMiddleware>[],
                                  middlewaresOrder: Type<RestMiddleware>[]): RestMiddlewareRunMethod[]
 {
     let usedMiddlewares: Type<RestMiddleware>[] = [];
@@ -31,7 +29,7 @@ export function buildMiddlewares(this: RESTClientInterface,
 
     usedMiddlewares.forEach(middleware =>
     {
-        runMethods.push(new middleware().run.bind(this));
+        runMethods.push(new middleware().run);
     });
 
     return runMethods;